fix(google-ads): build report windows in the account's own time zone
Google evaluates segments.date in the Ads account's own time zone. Report
windows were cut from the server's UTC clock, so after 5pm Pacific a
"today" that hasn't started there was included and a real day dropped off
the start. Existing connections keep UTC until the next verify, which is
the same window they already have.

Stores customer.time_zone (nullable) at account selection, from Google's
answer, never the client. The reporting service owns the default 28-day
window and the 7-day pacing window so the dashboard and MCP tools stop
computing their own UTC ranges. Explicit startDate/endDate still win.
Unknown or missing zones fall back to UTC.

Item 7 from the Google Ads review batch.
