+++ drizzle/meta/_journal.json
@@ -323,6 +323,13 @@
       "when": 1788993740768,
       "tag": "0045_google_ads_connections",
       "breakpoints": true
+    },
+    {
+      "idx": 46,
+      "version": "6",
+      "when": 1789006586830,
+      "tag": "0046_charming_overlord",
+      "breakpoints": true
     }
   ]
 }
