+++ src/db/pg/google-ads.schema.ts
@@ -20,6 +20,7 @@ export const googleAdsConnections = pgTable(
     loginCustomerId: text("login_customer_id"),
     customerDescriptiveName: text("customer_descriptive_name"),
     currencyCode: text("currency_code"),
+    timeZone: text("time_zone"),
     connectedByUserId: text("connected_by_user_id").notNull(),
     googleAdsAccountId: text("google_ads_account_id").notNull(),
     connectedAccountEmail: text("connected_account_email"),
