+++ src/db/schema-parity.test.ts
@@ -12,6 +12,7 @@ import * as sqliteSam from "./sam.schema";
 import * as sqliteAuth from "./better-auth-schema";
 import * as sqliteBilling from "./billing.schema";
 import * as sqliteGa4 from "./ga4.schema";
+import * as sqliteGoogleAds from "./google-ads.schema";
 import * as sqliteGsc from "./gsc.schema";
 import * as sqliteTelemetry from "./telemetry.schema";
 import * as pgApp from "./pg/app.schema";
@@ -21,6 +22,7 @@ import * as pgSam from "./pg/sam.schema";
 import * as pgAuth from "./pg/better-auth-schema";
 import * as pgBilling from "./pg/billing.schema";
 import * as pgGa4 from "./pg/ga4.schema";
+import * as pgGoogleAds from "./pg/google-ads.schema";
 import * as pgGsc from "./pg/gsc.schema";
 import * as pgTelemetry from "./pg/telemetry.schema";
 
@@ -151,6 +153,7 @@ const sqliteAppTables = tablesFrom(
   sqliteSam,
   sqliteBilling,
   sqliteGa4,
+  sqliteGoogleAds,
   sqliteGsc,
   sqliteTelemetry,
 );
@@ -161,6 +164,7 @@ const pgAppTables = tablesFrom(
   pgSam,
   pgBilling,
   pgGa4,
+  pgGoogleAds,
   pgGsc,
   pgTelemetry,
 );
